update according to the suggestions of @guitargeek
