+++ bindings/pyroot/pythonizations/python/ROOT/pythonization/_roofit/__init__.py
@@ -1,0 +1,1 @@
+from ._rooabspdf import pythonize_rooabspdf
